# backend/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ContextRuleResponse:
    properties:
      createdAt:
        type: string
      flagId:
        type: string
      id:
        type: string
      keyword:
        type: string
      subflagId:
        type: string
      updatedAt:
        type: string
    type: object
  dto.CreateContextRuleRequest:
    properties:
      flagId:
        type: string
      keyword:
        type: string
      subflagId:
        type: string
    type: object
  dto.CreateEventRequest:
    properties:
      allDay:
        type: boolean
      endAt:
        type: string
      location:
        type: string
      startAt:
        type: string
      title:
        type: string
    type: object
  dto.CreateFlagRequest:
    properties:
      color:
        type: string
      name:
        type: string
      sortOrder:
        type: integer
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      requestId:
        type: string
    type: object
  dto.EventResponse:
    properties:
      allDay:
        type: boolean
      createdAt:
        type: string
      endAt:
        type: string
      id:
        type: string
      location:
        type: string
      sourceInboxItemId:
        type: string
      startAt:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  dto.FlagResponse:
    properties:
      color:
        type: string
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      sortOrder:
        type: integer
      updatedAt:
        type: string
    type: object
...
  dto.ListContextRulesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.ContextRuleResponse'
        type: array
      nextCursor:
        type: string
    type: object
  dto.ListEventsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.EventResponse'
        type: array
      nextCursor:
        type: string
    type: object
  dto.ListFlagsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.FlagResponse'
        type: array
      nextCursor:
        type: string
    type: object
...
  dto.UpdateContextRuleRequest:
    properties:
      flagId:
        type: string
      keyword:
        type: string
      subflagId:
        type: string
    type: object
  dto.UpdateEventRequest:
    properties:
      allDay:
        type: boolean
      endAt:
        type: string
      location:
        type: string
      startAt:
        type: string
      title:
        type: string
    type: object
  dto.UpdateFlagRequest:
    properties:
      color:
        type: string
      name:
        type: string
      sortOrder:
        type: integer
    type: object
...
info:
  contact: {}
  description: API do MVP Inbota.
  title: Inbota API
  version: "0.1"
paths:
...
  /v1/context-rules:
    get:
      parameters:
      - description: Limite
        in: query
        name: limit
        type: integer
      - description: Cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListContextRulesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar regras de contexto
      tags:
      - ContextRules
    post:
      consumes:
      - application/json
      parameters:
      - description: Context rule payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateContextRuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ContextRuleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar regra de contexto
      tags:
      - ContextRules
  /v1/context-rules/{id}:
    delete:
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remover regra de contexto
      tags:
      - ContextRules
    patch:
      consumes:
      - application/json
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      - description: Context rule payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateContextRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ContextRuleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar regra de contexto
      tags:
      - ContextRules
  /v1/events:
    get:
      parameters:
      - description: Limite
        in: query
        name: limit
        type: integer
      - description: Cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListEventsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar eventos
      tags:
      - Events
    post:
      consumes:
      - application/json
      parameters:
      - description: Event payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar evento
      tags:
      - Events
  /v1/events/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Event payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar evento
      tags:
      - Events
  /v1/flags:
    get:
      parameters:
      - description: Limite
        in: query
        name: limit
        type: integer
      - description: Cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListFlagsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar flags
      tags:
      - Flags
    post:
      consumes:
      - application/json
      parameters:
      - description: Flag payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFlagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.FlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar flag
      tags:
      - Flags
  /v1/flags/{id}:
    delete:
      parameters:
      - description: Flag ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remover flag
      tags:
      - Flags
    patch:
      consumes:
      - application/json
      parameters:
      - description: Flag ID
        in: path
        name: id
        required: true
        type: string
      - description: Flag payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateFlagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar flag
      tags:
      - Flags
...
securityDefinitions:
  BearerAuth:
    description: 'Formato: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"